don't do thumbnail shooting on iconified or shaded windows.
Thu Mar  1 05:21:04 2001  Tim Janik  <timj@gtk.org>

        * gwmdesktop.c (thumb_queue_step): don't do thumbnail shooting on
        iconified or shaded windows.

        * gwmh.h (struct _GwmhTask): made width/height/x/y 32bit ints,
        to avoid crashes with excessive window sizes.

        * gwmdesktop.c (task_remove_thumb): fixed thumb_queue cleanup
        code (a bug that could cause segfaults after thumbnailing was
        turned back on).

        * gwmthumbnail.c (gwm_thumb_nail_destroy): pass the thumbnail in the
        destroy function.

svn path=/trunk/; revision=6611

+++ applets/desk-guide/gwmh.h
@@ -194,11 +194,11 @@ struct _GwmhTask
   guint		   iconified : 1;
   
   /* frame geometry */
-  gint16     frame_x, frame_y;
-  gint16     frame_width, frame_height;
+  gint       frame_x, frame_y;
+  gint       frame_width, frame_height;
   /* window geometry (within frame) */
-  gint16     win_x, win_y;
-  gint16     win_width, win_height;
+  gint       win_x, win_y;
+  gint       win_width, win_height;
 
   /* desktop, desktop area, layer */
   guint16    desktop;
